requirements.txt
langchain==0.1.6
langchain_community==0.0.19
langchain_core==0.1.22
langchain_openai==0.0.5
streamlit==1.31.0
python-dotenv
